// src/app/models/apimodel.ts
export interface Root {
    lat: number
    lon: number
    timezone: string
    timezone_offset: number
    current: Current
    daily: Daily[]
}

export interface Current {
    dt: number
    sunrise: number
    sunset: number
    temp: number
    feels_like: number
    pressure: number
    humidity: number
    dew_point: number
    uvi: number
    clouds: number
    visibility: number
    wind_speed: number
    wind_deg: number
    weather: Weather[]
}
  
export interface Weather {
    id: number
    main: string
    description: string
    icon: string
}
  
export interface Daily {
    dt: number
    sunrise: number
    sunset: number
    moonrise: number
    moonset: number
    moon_phase: number
    temp: Temp
    feels_like: FeelsLike
    pressure: number
    humidity: number
    dew_point: number
    wind_speed: number
    wind_deg: number
    wind_gust: number
    weather: Weather[]
    clouds: number
    pop: number
    uvi: number
    rain?: number
}
  
export interface Temp {
    day: number
    min: number
    max: number
    night: number
    eve: number
    morn: number
}
  
export interface FeelsLike {
    day: number
    night: number
    eve: number
    morn: number
}

export class Tempo {
    temp!: number;
    date!: string;
    time!: string;
    description!: string;
    city!: String;
    region!: string;
    icon!: string;
    forecast!: Forecast[];
    max!: number;
    min!: number;
}
  
export class Forecast {
    constructor(date: string, weekday: string, max: number, min: number, icon: string) {
        this.date = date;
        this.weekday = weekday;
        this.max = max;
        this.min = min;
        this.icon = icon;
    }

    date!: string
    weekday!: string
    max!: number
    min!: number
    icon!: string
}


export interface Cidade {
    name: String,
    lat: number,
    lon: number,
    country: string,
    state: string
}